Add files via upload
initial word_clouds.py commit. File paths will not work as of now. Will clone repo and change so that it will run in our file-structure and work with our dataset as they are in here. Three functions are defined in this script-we should discuss where we want to keep our functions for re-use